pyproject.toml
[project]
name = "mul-agent"
version = "0.1.0"
description = "Multi-Agent Collaboration Framework with Self-Evolution"
requires-python = ">=3.11"
dependencies = [
    # === 编排核心 ===
    "langgraph>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "langchain-openai>=0.3",
    "langchain-core>=0.3",

    # === Agent运行时 ===
    "openclaw-sdk",

    # === API网关 ===
    "fastapi>=0.115",
    "uvicorn[standard]",
    "sse-starlette",

    # === 存储 ===
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "qdrant-client>=1.12",
    "redis>=5.0",

    # === 安全审查 ===
    "docker>=7.0",

    # === 工具 ===
    "httpx",

    # === 基础 ===
    "pydantic>=2.0",
    "pydantic-settings",
    "pyyaml",
    "structlog",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx",
    "httpx",
    "ruff",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
